requirements.txt
google-genai>=1.64.0
jupyter>=1.1.0
notebook>=7.0.0
numpy>=1.24.0
pandas>=2.0.0